backend/requirement.txt
uvicorn
fastapi
aiosqlite
sqlmodel
pydantic
pyjwt
passlib[bcrypt]